@@ -27,22 +27,28 @@
     },
     "require": {
         "php": ">=8.0",
-        "symfony/ux-twig-component": "^2.0"
+        "symfony/ux-twig-component": "^2.1"
     },
     "require-dev": {
-        "symfony/form": "^4.4|^5.0|^6.0",
-        "symfony/framework-bundle": "^4.4|^5.0|^6.0",
-        "symfony/dependency-injection": "^4.4.2|^5.0|^6.0",
-        "symfony/twig-bundle": "^4.4|^5.0|^6.0",
-        "symfony/security-csrf": "^4.4|^5.0|^6.0",
-        "symfony/serializer": "^4.4|^5.0|^6.0",
-        "symfony/validator": "^4.4|^5.0|^6.0",
-        "symfony/phpunit-bridge": "^5.2|^6.0",
         "doctrine/annotations": "^1.0",
         "doctrine/doctrine-bundle": "^2.0",
         "doctrine/orm": "^2.7",
-        "zenstruck/foundry": "^1.10",
-        "zenstruck/browser": "^0.9.1"
+        "symfony/dependency-injection": "^5.4|^6.0",
+        "symfony/form": "^5.4|^6.0",
+        "symfony/framework-bundle": "^5.4|^6.0",
+        "symfony/phpunit-bridge": "^6.0",
+        "symfony/security-csrf": "^5.4|^6.0",
+        "symfony/serializer": "^5.4|^6.0",
+        "symfony/twig-bundle": "^5.4|^6.0",
+        "symfony/validator": "^5.4|^6.0",
+        "zenstruck/browser": "^0.9.1",
+        "zenstruck/foundry": "^1.10"
+    },
+    "conflict": {
+        "symfony/config": "<5.4.0"
+    },
+    "config": {
+        "sort-packages": true
     },
     "extra": {
         "thanks": {
