@@ -16,11 +16,10 @@ use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
 use Symfony\UX\LiveComponent\ComponentValidator;
 use Symfony\UX\LiveComponent\ComponentValidatorInterface;
+use Symfony\UX\LiveComponent\EventListener\AddLiveAttributesSubscriber;
 use Symfony\UX\LiveComponent\EventListener\LiveComponentSubscriber;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
 use Symfony\UX\LiveComponent\PropertyHydratorInterface;
@@ -38,21 +37,20 @@ final class LiveComponentExtension extends Extension
 {
     public function load(array $configs, ContainerBuilder $container): void
     {
-        if (method_exists($container, 'registerAttributeForAutoconfiguration')) {
-            $container->registerAttributeForAutoconfiguration(
-                AsLiveComponent::class,
-                function (ChildDefinition $definition, AsLiveComponent $attribute) {
-                    $definition
-                        ->addTag('twig.component', array_filter([
-                            'key' => $attribute->name,
-                            'template' => $attribute->template,
-                            'default_action' => $attribute->defaultAction,
-                        ]))
-                        ->addTag('controller.service_arguments')
-                    ;
-                }
-            );
-        }
+        $container->registerAttributeForAutoconfiguration(
+            AsLiveComponent::class,
+            function (ChildDefinition $definition, AsLiveComponent $attribute) {
+                $definition
+                    ->addTag('twig.component', array_filter([
+                        'key' => $attribute->name,
+                        'template' => $attribute->template,
+                        'default_action' => $attribute->defaultAction,
+                        'live' => true,
+                    ]))
+                    ->addTag('controller.service_arguments')
+                ;
+            }
+        );
 
         $container->registerForAutoconfiguration(PropertyHydratorInterface::class)
             ->addTag('twig.component.property_hydrator')
@@ -71,7 +69,7 @@ final class LiveComponentExtension extends Extension
             ->addTag('container.service_subscriber', ['key' => ComponentFactory::class, 'id' => 'ux.twig_component.component_factory'])
             ->addTag('container.service_subscriber', ['key' => ComponentRenderer::class, 'id' => 'ux.twig_component.component_renderer'])
             ->addTag('container.service_subscriber', ['key' => LiveComponentHydrator::class, 'id' => 'ux.live_component.component_hydrator'])
-            ->addTag('container.service_subscriber')
+            ->addTag('container.service_subscriber') // csrf
         ;
 
         $container->register('ux.live_component.twig.component_extension', LiveComponentTwigExtension::class)
@@ -82,14 +80,19 @@ final class LiveComponentExtension extends Extension
             ->setArguments([
                 new Reference('ux.live_component.component_hydrator'),
                 new Reference('ux.twig_component.component_factory'),
-                new Reference(UrlGeneratorInterface::class),
-                new Reference(CsrfTokenManagerInterface::class, ContainerBuilder::NULL_ON_INVALID_REFERENCE),
+                new Reference('router'),
             ])
             ->addTag('twig.runtime')
         ;
 
         $container->register(ComponentValidator::class)
             ->addTag('container.service_subscriber', ['key' => 'validator', 'id' => 'validator'])
+        ;
+
+        $container->register('ux.live_component.add_attributes_subscriber', AddLiveAttributesSubscriber::class)
+            ->addTag('kernel.event_subscriber')
+            ->addTag('container.service_subscriber', ['key' => LiveComponentHydrator::class, 'id' => 'ux.live_component.component_hydrator'])
+            ->addTag('container.service_subscriber') // csrf, twig & router
         ;
 
         $container->setAlias(ComponentValidatorInterface::class, ComponentValidator::class);
