@@ -14,9 +14,12 @@ declare(strict_types=1);
 namespace Symfony\UX\LiveComponent\Tests\Functional\EventListener;
 
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
+use Symfony\Component\DomCrawler\Crawler;
+use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
-use Symfony\UX\LiveComponent\Tests\ContainerBC;
 use Symfony\UX\LiveComponent\Tests\Fixture\Component\FormComponent1;
+use Symfony\UX\LiveComponent\Tests\Fixtures\Component\FormWithCollectionTypeComponent;
+use Symfony\UX\LiveComponent\Tests\Fixtures\Form\BlogPostFormType;
 use Symfony\UX\TwigComponent\ComponentFactory;
 use Zenstruck\Browser\Response\HtmlResponse;
 use Zenstruck\Browser\Test\HasBrowser;
@@ -28,11 +31,119 @@ use Zenstruck\Foundry\Test\ResetDatabase;
  */
 class ComponentWithFormTest extends KernelTestCase
 {
-    use ContainerBC;
     use Factories;
     use HasBrowser;
     use ResetDatabase;
 
+    public function testFormValuesRebuildAfterFormChanges(): void
+    {
+        /** @var LiveComponentHydrator $hydrator */
+        $hydrator = self::getContainer()->get('ux.live_component.component_hydrator');
+        /** @var ComponentFactory $factory */
+        $factory = self::getContainer()->get('ux.twig_component.component_factory');
+        $component = $factory->create('form_with_collection_type');
+
+        $dehydrated = $hydrator->dehydrate($component);
+        $token = null;
+
+        $this->browser()
+            ->get('/_components/form_with_collection_type?data='.urlencode(json_encode($dehydrated)))
+            ->use(function (HtmlResponse $response) use (&$dehydrated, &$token) {
+                // mimic user typing
+                $dehydrated['blog_post_form']['content'] = 'changed description by user';
+                $dehydrated['validatedFields'] = ['blog_post_form.content'];
+                $token = $response->crawler()->filter('div')->first()->attr('data-live-csrf-value');
+            })
+
+            // post to action, which will add a new embedded comment
+            ->post('/_components/form_with_collection_type/addComment', [
+                'body' => json_encode($dehydrated),
+                'headers' => ['X-CSRF-TOKEN' => $token],
+            ])
+            ->assertStatus(422)
+            // look for original embedded form
+            ->assertContains('<textarea id="blog_post_form_comments_0_content"')
+            // look for new embedded form
+            ->assertContains('<textarea id="blog_post_form_comments_1_content"')
+            // changed text is still present
+            ->assertContains('changed description by user</textarea>')
+            // check that validation happened and stuck
+            ->assertContains('The content field is too short')
+            // make sure the title field did not suddenly become validated
+            ->assertNotContains('The title field should not be blank')
+            ->use(function (Crawler $crawler) use (&$dehydrated, &$token) {
+                $div = $crawler->filter('[data-controller="live"]');
+                $liveData = json_decode($div->attr('data-live-data-value'), true);
+                // make sure the 2nd collection type was initialized, that it didn't
+                // just "keep" the empty array that we set it to in the component
+                $this->assertEquals(
+                    [
+                        ['content' => ''],
+                        ['content' => ''],
+                    ],
+                    $liveData['blog_post_form']['comments']
+                );
+
+                // grab the latest live data
+                $dehydrated = $liveData;
+                // fake that this field was being validated
+                $dehydrated['validatedFields'][] = 'blog_post_form.0.comments.content';
+                $token = $div->attr('data-live-csrf-value');
+            })
+
+            // post to action, which will remove the original embedded comment
+            ->post('/_components/form_with_collection_type/removeComment?'.http_build_query(['args' => 'index=0']), [
+                'body' => json_encode($dehydrated),
+                'headers' => ['X-CSRF-TOKEN' => $token],
+            ])
+            ->assertStatus(422)
+            // the original embedded form should be gone
+            ->assertNotContains('<textarea id="blog_post_form_comments_0_content"')
+            // the added one should still be present
+            ->assertContains('<textarea id="blog_post_form_comments_1_content"')
+            ->use(function (Crawler $crawler) {
+                $div = $crawler->filter('[data-controller="live"]');
+                $liveData = json_decode($div->attr('data-live-data-value'), true);
+                // the embedded validated field should be gone, since its data is gone
+                $this->assertEquals(
+                    ['blog_post_form.content'],
+                    $liveData['validatedFields']
+                );
+            })
+        ;
+    }
+
+    public function testFormRemembersValidationFromInitialForm(): void
+    {
+        /** @var LiveComponentHydrator $hydrator */
+        $hydrator = self::getContainer()->get('ux.live_component.component_hydrator');
+        /** @var ComponentFactory $factory */
+        $factory = self::getContainer()->get('ux.twig_component.component_factory');
+        /** @var FormFactoryInterface $formFactory */
+        $formFactory = self::getContainer()->get('form.factory');
+
+        $form = $formFactory->create(BlogPostFormType::class);
+        $form->submit(['title' => '', 'content' => '']);
+        /** @var FormWithCollectionTypeComponent $component */
+        $component = $factory->create('form_with_collection_type', [
+            'form' => $form->createView(),
+        ]);
+
+        // component should recognize that it is already submitted
+        $this->assertTrue($component->isValidated);
+
+        $dehydrated = $hydrator->dehydrate($component);
+        $dehydrated['blog_post_form']['content'] = 'changed description';
+        $dehydrated['validatedFields'][] = 'blog_post_form.content';
+
+        $this->browser()
+            ->get('/_components/form_with_collection_type?data='.urlencode(json_encode($dehydrated)))
+            // normal validation happened
+            ->assertContains('The content field is too short')
+            // title is STILL validated as all fields should be validated
+            ->assertContains('The title field should not be blank')
+        ;
+    }
     public function testHandleCheckboxChanges(): void
     {
         /** @var LiveComponentHydrator $hydrator */
